requirements.txt
bcrypt==3.2.0
cffi==1.14.5
click==8.0.0
Flask==2.0.0
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-Login==0.5.0
itsdangerous==2.0.0
Jinja2==3.0.0
MarkupSafe==2.0.0
peewee==3.14.4
psycopg2-binary==2.8.6
pycparser==2.20
python-dotenv==0.17.1
six==1.16.0
Werkzeug==2.0.0
